Make it work in IE8 by default
This makes sure that certain runtime dependencies are always available on
the client. One of those being es5-shim.

I also kill the full document rendering bugfix since this is fixed in
master.

Since IE doesn't support DOMContentLoaded I just execute it immediately.
We may want to load scripts synchronously, in which case we'll need to
fix this. But since this code currently have to run synchronously, this
is the easy fix.