Allow source map breakpoints to be set on a line
This wasn't accurate since it skips a column.

Safari and JSC treats column 0 as special so there needs to be a mapping
for the first column to set breakpoints. Source maps are inherently broken
for breakpoints in minified code but this works for the kind of 1:1
mapping between lines we do here.